fix(prettyprint): support whitespace preservation
Whitespace is captured as Text node in AST. When such Text nodes for spaces appear as last tag in body of parent tag,
they get ignored due to trim. When a Text node contains a literal value with leading whitespace, whitespace is left trimmed.
This commint addresses both of these issues.